Adding message documentation.
Per Raymond's suggestion in change I0188900f, this change will add
message documentation strings so that it can be translated on
Translatewiki.net.

Note: I am open to suggestions to change the usage of the phrase
"Host statistics", as it can be quite confusing. It generally means
the information about the server.

Change-Id: I67fcda8a9048927ed0312eaed2e263a38c532265